TIP-843: Move product analyser to enrich
