+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/ProductAnalyzer.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Pim\Component\Connector\Analyzer;
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector;
 
 use Akeneo\Tool\Component\Batch\Item\ItemReaderInterface;
+use Pim\Component\Connector\Analyzer\AnalyzerInterface;
 
 /**
  * Provides some statistics on a products data provided by a reader
