+++ src/Pim/Bundle/ConnectorBundle/DependencyInjection/PimConnectorExtension.php
@@ -22,7 +22,6 @@ class PimConnectorExtension extends Extension
     public function load(array $configs, ContainerBuilder $container)
     {
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
-        $loader->load('analyzers.yml');
         $loader->load('archiving.yml');
         $loader->load('array_converters.yml');
         $loader->load('doctrine.yml');
