--- src/Pim/Bundle/ConnectorBundle/Resources/config/analyzers.yml
@@ -1,6 +1,0 @@
-parameters:
-    pim_connector.analyzer.product.class: Pim\Component\Connector\Analyzer\ProductAnalyzer
-
-services:
-    pim_connector.analyzer.product:
-        class: '%pim_connector.analyzer.product.class%'
